requirements.txt
# AutoGen 核心框架
pyautogen>=0.2.0
autogen-agentchat>=0.2.0

# LLM 集成
openai>=1.12.0
anthropic>=0.18.0

# 可观测性 - OpenTelemetry 生态
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-openai>=0.19.0
opentelemetry-instrumentation-anthropic>=0.19.0
arize-phoenix>=3.0.0

# 数据验证和安全
pydantic>=2.5.0
instructor>=0.4.0

# 内容安全
openai  # 用于 Moderation API

# 工具和实用程序
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
httpx>=0.25.0

# 开发和测试
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0 